Uniform sampling of the environmental space

This function allows to perform a uniform sampling of the environmental space. It requires an sf object having point geometry given by the PC-scores values (the first two axes of a PCA computed on the environmental variables of interest), and the bounding box of the environmental space (defined by the extent of the PC-scores values). A resolution of the sampling grid must be provided by the user. The resolution can be arbitrarily selected or or defined using the optimRes() function. The functions allows to retrieve both a training and a testing dataset (optional) of uniformly sampled points. The user must provide suitable numbers of points to be sampled in each cell of the sampling grid (n.tr: points for the training dataset; n.ts: points for the testing dataset). Notice that the final number of sampled background points depends on the PC-scores configuration in the environmental space.

uniformSampling(
  sdf,
  grid.res,
  n.tr = 5,
  n.prev = NULL,
  sub.ts = FALSE,
  n.ts = 5,
  plot_proc = FALSE,
  verbose = FALSE
)

Arguments

sdf

an sf object having point geometry given by the PC-scores values

grid.res

(integer) resolution of the sampling grid. The resolution can be arbitrarily selected or defined using the optimRes() function.

n.tr

(integer; optional) number of expected points given a certain prevalence threshold for the training dataset.

n.prev

(double) sample prevalence

sub.ts

(logical) sample the validation points

n.ts

(integer; optional) number of points for the testing dataset to sample in each cell of the sampling grid. sub.ts argument must be TRUE.

plot_proc

(logical) plot progress of the sampling

verbose

(logical) Print verbose

Value

A spatial point data frame with the coordinates of the sampled points both in the geographical and environmental space